G2P-760 G2P-790 :Create application some part of code changed to dswd, Enrollment date moved to the g2p_program
