+++ g2p_program_registrant_info/__manifest__.py
@@ -18,8 +18,6 @@
         "views/program_registrant_info.xml",
         "views/program_membership.xml",
         "security/ir.model.access.csv",
-        # wizard
-        "wizard/g2p_program_registrant_info_wizard.xml",
     ],
     "assets": {
         "web.assets_backend": [
