+++ g2p_program_registrant_info/wizard/g2p_program_registrant_info_wizard.py
@@ -4,7 +4,7 @@ import base64
 import logging
 from io import BytesIO
 
-from odoo import fields, models
+from odoo import models
 
 _logger = logging.getLogger(__name__)
 
@@ -12,14 +12,6 @@ _logger = logging.getLogger(__name__)
 class G2pProgramRegistrantInfo(models.TransientModel):
     _name = "g2p.program.registrantinfo.wizard"
     _description = "G2P Program Registrant Info Wizard"
-
-    reason_for_app = fields.Char(string="Reason for Application")
-    upload_doc1 = fields.Binary(string="Medical Prescription", attachment=True)
-    upload_doc2 = fields.Binary(string="Medical Diagnosis", attachment=True)
-    upload_doc3 = fields.Binary(
-        string="Additional Supporting Document", attachment=True
-    )
-    additional_info = fields.Char(string="Additional Information")
 
     def jsonize_form_data(self, data, program, membership=None):
         try:
@@ -73,40 +65,3 @@ class G2pProgramRegistrantInfo(models.TransientModel):
                 "An error occurred while adding files to the store: %s", str(e)
             )
         return file_details
-
-    def create_registrantinfo(self):
-        membership = (
-            self.env["g2p.program_membership"]
-            .sudo()
-            .search([("id", "=", self._context.get("active_id"))])
-        )
-        partners = membership.partner_id
-        program = membership.program_id
-
-        if not partners or not program:
-            return
-
-        try:
-            data = {
-                "Reason for Application": self.reason_for_app or None,
-                "Document1": self.upload_doc1 or None,
-                "Document2": self.upload_doc2 or None,
-                "Document3": self.upload_doc3 or None,
-                "Additional Information": self.additional_info or None,
-            }
-
-            processed_data = self.jsonize_form_data(data, program, membership[0])
-
-            self.env["g2p.program.registrant_info"].sudo().create(
-                {
-                    "state": "active",
-                    "program_registrant_info": processed_data,
-                    "program_id": program.id,
-                    "registrant_id": partners.id,
-                }
-            )
-
-        except Exception as e:
-            _logger.exception(
-                "An error occurred while creating registrant info: %s", str(e)
-            )
