--- g2p_program_registrant_info/wizard/g2p_program_registrant_info_wizard.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-Part of OpenG2P. See LICENSE file for full copyright and licensing details.
--->
-<odoo>
-    <record id="view_program_registrantInfo_wizard_form" model="ir.ui.view">
-        <field name="name">view_program_registrantinfo_wizard_form</field>
-        <field name="model">g2p.program.registrantinfo.wizard</field>
-        <field name="type">form</field>
-        <field name="arch" type="xml">
-            <form>
-                <sheet>
-                    <group>
-                        <field name="reason_for_app" required="1" />
-                        <field name="upload_doc1" />
-                        <field name="upload_doc2" />
-                        <field name="upload_doc3" />
-                        <field name="additional_info" />
-                    </group>
-                </sheet>
-                <footer>
-                    <button name="create_registrantinfo" type="object" string="Create" class="btn-primary" />
-                    <button string="Cancel" class="btn-secondary" special="cancel" />
-                </footer>
-            </form>
-        </field>
-    </record>
-</odoo>
