+++ g2p_programs/models/program_membership.py
@@ -1,7 +1,7 @@
 # Part of OpenG2P. See LICENSE file for full copyright and licensing details.
 from lxml import etree
 
-from odoo import _, fields, models
+from odoo import _, api, fields, models
 
 from . import constants
 
@@ -39,7 +39,8 @@ class G2PProgramMembership(models.Model):
         copy=False,
     )
 
-    enrollment_date = fields.Date(default=lambda self: fields.Datetime.now())
+    enrollment_date = fields.Date(compute="_compute_enrolled_date", store=True)
+
     last_deduplication = fields.Date("Last Deduplication Date")
     exit_date = fields.Date()
 
@@ -77,6 +78,14 @@ class G2PProgramMembership(models.Model):
         default="new",
         copy=False,
     )
+
+    @api.depends("state")
+    def _compute_enrolled_date(self):
+        for rec in self:
+            if rec.state == "enrolled":
+                rec.enrollment_date = fields.Datetime.now()
+            else:
+                rec.enrollment_date = None
 
     def fields_view_get(
         self, view_id=None, view_type="form", toolbar=False, submenu=False
